refactor course pages into page wrappers hooks and shared notifier
