+++ frontend/src/components/Course/Course/partials/AddCourseForm.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { Form } from "react-bootstrap";
+
+const AddCourseForm = ({ onSubmit, children }) => (
+  <Form onSubmit={onSubmit}>{children}</Form>
+);
+
+export default AddCourseForm;
